requirements.txt
torch
scikit-learn
pandas
datasets
gdown
ragatouille
colbert-ai
django 
celery 
redis 
python-dotenv 
